+++ game/request.go
@@ -45,16 +45,16 @@ func UnmarshalRequest(_b []byte) Request {
 func (r *Request) PayloadToPlayer() *Player {
 	p := r.Payload.(map[string]interface{})
 	return &Player{
-		int(p["Id"].(int64)),
+		int(p["Id"].(int8)),
 		p["Name"].(string),
 		p["Pos_x"].(float64),
 		p["Pos_y"].(float64),
 		p["Rotation"].(float64),
 		p["FOV"].(float64),
 		p["Hp"].(float64),
-		int(p["Ammo"].(int64)),
-		int(p["WeaponOwned"].(int64)),
-		PlayerState(p["State"].(int64)),
+		int(p["Ammo"].(int8)),
+		int(p["WeaponOwned"].(int8)),
+		PlayerState(p["State"].(int8)),
 	}
 }
 
@@ -64,10 +64,10 @@ func (r *Request) PayloadToBullet() Bullet {
 		int(p["Id"].(int64)),
 		int(p["Owner_id"].(int64)),
 		int(p["Bullet_type"].(int64)),
-		p["Pos_x"].(float64),
-		p["Pos_y"].(float64),
-		p["Rotation"].(float64),
-		p["Distance"].(float64),
+		float64(p["Pos_x"].(float64)),
+		float64(p["Pos_y"].(float64)),
+		float64(p["Rotation"].(float64)),
+		float64(p["Distance"].(float64)),
 	}
 }
 
@@ -76,8 +76,8 @@ func (r *Request) PayloadToWeaponDrop() WeaponDrop {
 	return WeaponDrop{
 		int(p["Id"].(int64)),
 		int(p["Type_id"].(int64)),
-		p["Pos_x"].(float64),
-		p["Pos_y"].(float64),
+		float64(p["Pos_x"].(float64)),
+		float64(p["Pos_y"].(float64)),
 		int(p["Owner_id"].(float64)),
 	}
 }
